Fix FPU compare code.
The original code from:

    https://opencores.org/projects/fpu

generates completely reversed results
when comparing floats. Someone submitted a bug
about this eight months ago:

    https://opencores.org/projects/fpu/issues/2

Supposedly the author of the code tested the
comparison routine with 600k test vectors, but
the test bench must also be backward. Clearly
this code isn't to be trusted. I've fixed it here
but we should find a better library.